move logic for batch changes credentials around
